+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Upcoming
 * Added Creeper Collateral option. Disabled by default. If enabled will make creepers' explosion drop no blocks.
+* Tiredness
+  * You can now sleep during daytime only if you have the Tired effect and not if you're tired
 * Updated effects icons
 * Food is now eaten slightly faster
 * Freeze no longer applies injured